fix: remove not-required logging within test cases.
fix prometheus summary on time cost.

+++ diskcache/get.go
@@ -74,7 +74,7 @@ func (c *DiskCache) doGet(buf *bytes.Buffer, fn Fn) error {
 			getBytesVec.WithLabelValues(c.path).Observe(float64(nbytes))
 
 			// get on EOF not counted as a real Get
-			getLatencyVec.WithLabelValues(c.path).Observe(float64(time.Since(start) / time.Second))
+			getLatencyVec.WithLabelValues(c.path).Observe(float64(time.Since(start)) / float64(time.Second))
 		}
 	}()
